# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Task": {
				"properties": {
					"ofCalendarId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"createdById": {
						"type": "number",
						"format": "double"
					},
					"isGlobal": {
						"type": "boolean",
						"nullable": true
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ofCalendarId",
					"createdById",
					"isGlobal",
					"endDate",
					"startDate",
					"description",
					"type",
					"id"
				],
				"type": "object",
				"description": "Model Task"
			},
			"CalendarResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"tasks": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Calendar": {
				"properties": {
					"createdById": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"createdById",
					"description",
					"name",
					"id"
				],
				"type": "object",
				"description": "Model Calendar"
			},
			"CalendarUpdateParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"user": {
				"properties": {
					"passwordResetExpireDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"passwordResetCode": {
						"type": "string",
						"nullable": true
					},
					"twitterId": {
						"type": "string",
						"nullable": true
					},
					"facebookId": {
						"type": "string",
						"nullable": true
					},
					"googleId": {
						"type": "string",
						"nullable": true
					},
					"profilePhoto": {
						"type": "string",
						"nullable": true
					},
					"role": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"nominative": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"passwordResetExpireDate",
					"passwordResetCode",
					"twitterId",
					"facebookId",
					"googleId",
					"profilePhoto",
					"role",
					"password",
					"email",
					"nominative",
					"id"
				],
				"type": "object",
				"description": "Model user"
			},
			"TaskResponse": {
				"properties": {
					"type": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"isGlobal": {
						"type": "boolean"
					},
					"createdBy": {
						"$ref": "#/components/schemas/user"
					}
				},
				"required": [
					"type",
					"description",
					"startDate",
					"endDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TaskCreateParams": {
				"properties": {
					"type": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"isGlobal": {
						"type": "boolean"
					},
					"calendarId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"type",
					"description",
					"startDate",
					"endDate",
					"isGlobal",
					"calendarId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TaskUpdateParams": {
				"properties": {
					"type": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"isGlobal": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AuthResponse": {
				"properties": {
					"token": {
						"additionalProperties": false,
						"type": "object"
					},
					"user": {},
					"error": {}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ForgotPasswordParams": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SignupParams": {
				"properties": {
					"nominative": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"profilePhoto": {
						"type": "string"
					}
				},
				"required": [
					"nominative",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateParams": {
				"properties": {
					"nominative": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"profilePhoto": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "project-c-backend",
		"version": "0.0.1",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/calendar": {
			"post": {
				"operationId": "CreateCalendar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CalendarResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Calendar"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateParams"
							}
						}
					}
				}
			}
		},
		"/calendar/user/{userId}": {
			"get": {
				"operationId": "GetUserCalendars",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CalendarResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Calendar"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/calendar/{id}": {
			"get": {
				"operationId": "GetCalendar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CalendarResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Calendar"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateCalendar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Calendar"
								}
							}
						}
					}
				},
				"tags": [
					"Calendar"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CalendarUpdateParams"
							}
						}
					}
				}
			}
		},
		"/task": {
			"post": {
				"operationId": "CreateTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Task"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TaskCreateParams"
							}
						}
					}
				}
			}
		},
		"/task/user/{userId}": {
			"get": {
				"operationId": "GetUserTasks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TaskResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Task"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/task/calendar/{calendarId}": {
			"get": {
				"operationId": "GetCalendarTasks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TaskResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Task"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "calendarId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/task/{id}": {
			"get": {
				"operationId": "GetTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Task"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Task"
								}
							}
						}
					}
				},
				"tags": [
					"Task"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TaskUpdateParams"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "LocalLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/auth/signup": {
			"post": {
				"operationId": "LocalSignup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/auth/forgotPassword": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ForgotPasswordParams"
							}
						}
					}
				}
			}
		},
		"/auth/verifyPasswordResetCode": {
			"post": {
				"operationId": "VerifyPasswordResetCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"passwordResetCode": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"passwordResetCode",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/updatePassword": {
			"post": {
				"operationId": "UpdatePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"newPassword": {
										"type": "string"
									},
									"passwordResetCode": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"newPassword",
									"passwordResetCode",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/user"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignupParams"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/user"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": []
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/user"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/user"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateParams"
							}
						}
					}
				}
			}
		},
		"/users/task/{taskId}": {
			"get": {
				"operationId": "GetTaskCreator",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/user"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}